+++ src/nlp_shap/value/__init__.py
@@ -1,0 +1,12 @@
+"""Coalition utility scoring plugins."""
+
+from .embedding import CosineEmbeddingValue, EuclideanEmbeddingValue
+from .logprob import LogprobValue
+from .tfidf import TfIdfCosineValue
+
+__all__ = [
+    "CosineEmbeddingValue",
+    "EuclideanEmbeddingValue",
+    "LogprobValue",
+    "TfIdfCosineValue",
+]